Add Version Badge
Version Badge provides a consistent way for the Ruby community to learn about the RubyGem associated with a particular Github repository and other documentation pages. Once the Gem owner adds this badge to their README file, it will inform and link all visitors to the latest version of that Gem.